registration and logout done

+++ algoscale/login/views.py
@@ -7,6 +7,16 @@ from django.template.context_processors import csrf
 from django.contrib import auth
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from .forms import RegistrationForm
+from django.template import RequestContext
+
+
+def index(request):
+    if request.user.is_authenticated:
+        return HttpResponseRedirect('/profile')
+    context = {}
+    context.update(csrf(request))
+    return render(request, 'login.html', context)
 
 
 def user_login(request):
@@ -27,29 +37,68 @@ def auth_view(request):
         print(request.user.username)
         return HttpResponseRedirect('/profile')
     else:
-        msg = "Wrong Username or Password ....Please Try again."
-        return render(request, 'login.html', {'msg': msg})
+        # args = dict()
+        # args["err"] = "Login failed...try registering yourself"
+        # args["form"] = RegistrationForm()
+        # return render(request, 'register.html', args)
+        return HttpResponseRedirect('/user_registration')
 
 
-@login_required(login_url='/')
+@login_required(login_url='/login')
 def profile(request):
-    context = {}
+    login_user_id = request.user.id
+    print("login_user_id ::", login_user_id)
+
+    if request.method == "POST":
+        try:
+            user_id = request.POST.get('user_id', "")
+            user_obj = User.objects.get(id=user_id)
+            user_obj.delete()
+            msg = "Successfully deleted user with ID {0}".format(user_id)
+        except:
+            if user_id == "":
+                msg = "Blank user ID".format(user_id)
+            else:
+                msg = "Unable to delete user with ID {0}".format(user_id)
+        context = {'msg': msg}
+
+    else:
+        context = {}
+
     users = User.objects.all()
+    users = users.exclude(id=login_user_id)
     context['users'] = users
+    if len(users) == 0:
+        context['msg'] = "No user to delete. You are the only user in this system"
     context.update(csrf(request))
     return render(request, 'profile.html', context=context)
 
 
-def delete_user(request):
-    try:
-        user_id = request.POST.get('user_id', "")
-        user_obj = User.objects.get(id=user_id)
-        user_obj.delete()
-        msg = "Successfully deleted user with ID {0}".format(user_id)
-    except:
-        if user_id == "":
-            msg = "Blank user ID".format(user_id)
-        else:
-            msg = "Unable to delete user with ID {0}".format(user_id)
+def logout(request):
+    auth.logout(request)
+    msg = "Successfully Logout...."
+    return render(request, 'login.html', {"msg": msg})
 
-    return render(request, 'profile.html', {'msg': msg})
+
+def user_registration(request):
+    if request.user.is_authenticated:
+        return HttpResponseRedirect('/profile')
+    args = {}
+    args.update(csrf(request))
+    if request.method == "POST":
+        print("post request")
+        form = RegistrationForm(request.POST)
+        args["err"] = form.errors
+        if form.is_valid():
+            user = form.save()
+            return HttpResponseRedirect('/register_success')
+
+        args["form"] = form
+    else:
+        args["form"] = RegistrationForm()
+    return render(request, 'register.html', args)
+
+
+def register_success(request):
+    msg = "Register Successful ...Please Login"
+    return render(request, 'login.html', {'msg': msg})
